merge-main (#521)
* port fallback for tendermint (#507)

* test on fedora (#1)

* always run (#2)

* implement port fallback

* test entire app

* don'pt start tendermint

* autostart

* simplify

* make it work

* reset github workflows

* appease linter

* appease linter 2

* start checking ports at 8080

* let docker choose port

* pick port in run.py

* set FLASK_PORT in run.py

* cleaner

* chain the exception

* simpler port selection

* make closer to original

* closer to original

* fix port

* use bridge mode on linux

* Revert "port fallback for tendermint (#507)"

This reverts commit 7afc6c3f8fb2bfe324438efa96371cfd356f0ba9.

* Feat/run service (#512)

* feat/moved-to-services

* refactor:moved-to-new-service

* refactor:implemented-base-class

* refactor:implemented-simple-tests

* feat:service-attempts-to-start

* feat/moved-to-services

* fix:run.py

---------


---------
