+++ auto_dev/commands/publish.py
@@ -1,13 +1,12 @@
 """This module contains the logic for the publish command."""
 
-
 import rich_click as click
 from aea.configurations.base import PublicId
 
 from auto_dev.base import build_cli
 from auto_dev.constants import AGENT_PUBLISHED_SUCCESS_MSG
 from auto_dev.exceptions import OperationError
-from auto_dev.commands.run import AgentRunner
+from auto_dev.commands.run import DevAgentRunner
 from auto_dev.services.package_manager.index import PackageManager
 
 
@@ -46,7 +45,7 @@ def publish(ctx, public_id: PublicId = None, force: bool = False) -> None:
     logger = ctx.obj["LOGGER"]
 
     try:
-        agent_runner = AgentRunner(
+        agent_runner = DevAgentRunner(
             agent_name=public_id,
             logger=logger,
             verbose=verbose,
@@ -54,9 +53,7 @@ def publish(ctx, public_id: PublicId = None, force: bool = False) -> None:
         )
         if not agent_runner.is_in_agent_dir():
             msg = "Not in an agent directory (aea-config.yaml not found) Please enter the agent directory to publish"
-            raise OperationError(
-                msg
-            )
+            raise OperationError(msg)
         package_manager = PackageManager(verbose=verbose)
         package_manager.publish_agent(force=force, new_public_id=public_id)
         click.secho(AGENT_PUBLISHED_SUCCESS_MSG, fg="green")
