+++ auto_dev/commands/run.py
@@ -1,30 +1,15 @@
 """Command to run an agent."""
 
-import os
 import sys
-import time
-import shutil
-import platform
-import subprocess
-from copy import deepcopy
-from typing import Any
 from pathlib import Path
-from textwrap import dedent
-from dataclasses import dataclass
 
-import docker
-import requests
 import rich_click as click
-from docker.errors import NotFound
 from aea.skills.base import PublicId
 from aea.configurations.base import PackageType
-from aea.configurations.constants import DEFAULT_AEA_CONFIG_FILE
 
 from auto_dev.base import build_cli
-from auto_dev.utils import change_dir, map_os_to_env_vars, load_autonolas_yaml
-from auto_dev.constants import DOCKERCOMPOSE_TEMPLATE_FOLDER
-from auto_dev.exceptions import UserInputError
-from auto_dev.cli_executor import CommandExecutor
+from auto_dev.utils import load_autonolas_yaml
+from auto_dev.services.runner import DevAgentRunner, ProdAgentRunner
 
 
 TENDERMINT_RESET_TIMEOUT = 10
@@ -34,278 +19,12 @@ TENDERMINT_RESET_RETRIES = 20
 cli = build_cli()
 
 
-@dataclass
-class AgentRunner:
-    """Class to manage running an agent."""
-
-    agent_name: PublicId
-    verbose: bool
-    force: bool
-    logger: Any
-    fetch: bool = False
-
-    def run(self) -> None:
-        """Run the agent."""
-        agent_path = "." if not self.fetch else self.agent_name.name
-        if self.fetch:
-            self.fetch_agent()
-        if not self.check_agent_exists(locally=True, in_packages=False):
-            self.logger.error(f"Local agent package {self.agent_name.name} does not exist.")
-            sys.exit(1)
-        self.logger.info(f"Changing to directory: {agent_path}")
-        with change_dir(agent_path):
-            self.check_tendermint()
-            self.setup_agent()
-            self.execute_agent()
-        self.stop_tendermint()
-
-    def check_agent_exists(self, locally=False, in_packages=True) -> bool:
-        """Check if the agent exists."""
-
-        if locally and in_packages:
-            msg = "Cannot check both locally and in packages."
-            raise UserInputError(msg)
-        if locally:
-            return self._is_locally_fetched() or self.is_in_agent_dir()
-        if in_packages:
-            return self._is_in_packages()
-        return False
-
-    def _is_locally_fetched(self):
-        return Path(self.agent_name.name).exists()
-
-    def is_in_agent_dir(self):
-        """Check if the agent is in the agent directory."""
-        return Path(DEFAULT_AEA_CONFIG_FILE).exists()
-
-    def _is_in_packages(self):
-        return Path(self.agent_package_path).exists()
-
-    @property
-    def agent_package_path(self):
-        """Get the agent package path."""
-        return Path("packages") / self.agent_name.author / "agents" / self.agent_name.name
-
-    @property
-    def agent_dir(self) -> Path:
-        """Get the agent directory based on where it is found."""
-        if self.is_in_agent_dir():
-            return Path()
-        if self._is_locally_fetched():
-            return Path(self.agent_name.name)
-        if self._is_in_packages():
-            return self.agent_package_path
-        msg = f"Agent not found. {self.agent_name} not found in local packages or agent directory."
-        raise UserInputError(msg)
-
-    def stop_tendermint(self) -> None:
-        """Stop Tendermint."""
-        self.execute_command(f"docker compose -f {DOCKERCOMPOSE_TEMPLATE_FOLDER}/tendermint.yaml kill")
-        self.execute_command(f"docker compose -f {DOCKERCOMPOSE_TEMPLATE_FOLDER}/tendermint.yaml down")
-        self.logger.info("Tendermint stopped. 🛑")
-
-    def check_tendermint(self, retries: int = 0) -> None:
-        """Check if Tendermint is running."""
-        self.logger.info("Checking Tendermint status...")
-        docker_engine = docker.from_env()
-        os_name = platform.system()
-        container_name = "tm_0"
-        tm_overrides = map_os_to_env_vars(os_name)
-        try:
-            self.logger.debug(f"Looking for Tendermint container: {container_name}")
-            res = docker_engine.containers.get(container_name)
-            self.logger.info(f"Found Tendermint container with status: {res.status}")
-            if res.status == "exited":
-                res.remove()
-                time.sleep(0.2)
-                self.check_tendermint(retries + 1)
-            if res.status == "running":
-                self.attempt_hard_reset()
-        except (subprocess.CalledProcessError, RuntimeError, NotFound) as e:
-            self.logger.info(f"Tendermint container not found or error: {e}")
-            if retries > 3:
-                self.logger.exception(f"Tendermint is not running. Please install and run Tendermint using Docker. {e}")
-                sys.exit(1)
-            self.logger.info("Starting Tendermint... 🚀")
-            self.start_tendermint(tm_overrides)
-            time.sleep(2)
-            return self.check_tendermint(retries + 1)
-        if res.status != "running":
-            self.logger.error("Tendermint is not healthy. Please check the logs.")
-            sys.exit(1)
-
-        self.logger.info("Tendermint is running and healthy ✅")
-        return None
-
-    def attempt_hard_reset(self, attempts: int = 0) -> None:
-        """Attempt to hard reset Tendermint."""
-        if attempts >= TENDERMINT_RESET_RETRIES:
-            self.logger.error(f"Failed to reset Tendermint after {TENDERMINT_RESET_RETRIES} attempts.")
-            sys.exit(1)
-
-        self.logger.info("Tendermint is running, executing hard reset...")
-        try:
-            response = requests.get(TENDERMINT_RESET_ENDPOINT, timeout=TENDERMINT_RESET_TIMEOUT)
-            if response.status_code == 200:
-                self.logger.info("Tendermint hard reset successful.")
-                return
-        except requests.RequestException as e:
-            self.logger.info(f"Failed to execute hard reset: {e}")
-
-        self.logger.info(f"Tendermint not ready (attempt {attempts + 1}/{TENDERMINT_RESET_RETRIES}), waiting...")
-        time.sleep(1)
-        self.attempt_hard_reset(attempts + 1)
-
-    def fetch_agent(self) -> None:
-        """Fetch the agent from registry if needed."""
-        self.logger.info(f"Fetching agent {self.agent_name} from the local package registry...")
-
-        if self.check_agent_exists(locally=True, in_packages=False):
-            if not self.force:
-                self.logger.error(f"Agent `{self.agent_name}` already exists. Use --force to overwrite.")
-                sys.exit(1)
-            self.logger.info(f"Removing existing agent `{self.agent_name}` due to --force option.")
-            self.execute_command(f"rm -rf {self.agent_name.name}")
-
-        command = f"aea -s fetch {self.agent_name} --local"
-        if not self.execute_command(command):
-            self.logger.error(f"Failed to fetch agent {self.agent_name}.")
-            sys.exit(1)
-
-    def setup_agent(self) -> None:
-        """Setup the agent."""
-        if not self.fetch:
-            self.logger.info(f"Agent author: {self.agent_name.author}")
-            self.logger.info(f"Agent name: {self.agent_name.name}")
-
-        self.logger.info("Setting up agent keys...")
-        self.manage_keys()
-
-        self.logger.info("Installing dependencies...")
-        self.install_dependencies()
-
-        self.logger.info("Setting up certificates...")
-        self.issue_certificates()
-        self.logger.info("Agent setup complete. 🎉")
-
-    def manage_keys(
-        self,
-        generate_keys: bool = True,
-    ) -> None:
-        """Manage keys based on the agent's default ledger configuration."""
-        config = load_autonolas_yaml(PackageType.AGENT)[0]
-        required_ledgers = config["required_ledgers"]
-        if not required_ledgers:
-            self.logger.error("No ledgers found in the agent configuration.")
-            sys.exit(1)
-        for ledger in required_ledgers:
-            self.logger.info(f"Processing ledger: {ledger}")
-            self.setup_ledger_key(ledger, generate_keys)
-
-    def setup_ledger_key(self, ledger: str, generate_keys, existing_key_file: Path | None = None) -> None:
-        """Setup the agent with the ledger key."""
-        key_file = Path(f"{ledger}_private_key.txt")
-        commands_to_errors = []
-        if existing_key_file:
-            self.logger.info(f"Copying existing key file {existing_key_file} to {key_file}")
-            shutil.copy(existing_key_file, key_file)
-        if key_file.exists():
-            self.logger.error(f"Key file {key_file} already exists.")
-        else:
-            if generate_keys:
-                self.logger.info(f"Generating key for {ledger}...")
-                commands_to_errors.append([f"aea -s generate-key {ledger}", f"Key generation failed for {ledger}"])
-            commands_to_errors.append([f"aea -s add-key {ledger}", f"Key addition failed for {ledger}"])
-
-        for command, error in commands_to_errors:
-            result = self.execute_command(command)
-            if not result:
-                self.logger.error(error)
-        self.logger.info(f"{ledger} key setup complete ✅")
-
-    def install_dependencies(self) -> None:
-        """Install agent dependencies."""
-        self.execute_command("aea -s install")
-
-    def issue_certificates(self) -> None:
-        """Issue certificates for agent if needed."""
-        if not Path("../certs").exists():
-            self.execute_command("aea -s issue-certificates")
-        else:
-            self.execute_command("cp -r ../certs ./")
-
-    def start_tendermint(self, env_vars=None) -> None:
-        """Start Tendermint."""
-        self.logger.info("Starting Tendermint with docker-compose...")
-        try:
-            result = self.execute_command(
-                f"docker compose -f {DOCKERCOMPOSE_TEMPLATE_FOLDER}/tendermint.yaml up -d --force-recreate",
-                env_vars=env_vars,
-            )
-            if not result:
-                msg = "Docker compose command failed to start Tendermint"
-                raise RuntimeError(msg)
-            self.logger.info("Tendermint started successfully")
-        except FileNotFoundError:
-            self.logger.exception("Docker compose file not found. Please ensure Tendermint configuration exists.")
-            sys.exit(1)
-        except docker.errors.DockerException as e:
-            self.logger.exception(
-                f"Docker error: {e!s}. Please ensure Docker is running and you have necessary permissions."
-            )
-            sys.exit(1)
-        except Exception as e:
-            self.logger.exception(f"Failed to start Tendermint: {e!s}")
-
-            msg = dedent("""
-                         Please check that:
-                         1. Docker is installed and running
-                         2. Docker compose is installed
-                         3. You have necessary permissions to run Docker commands
-                         4. The Tendermint configuration file exists and is valid
-                         """)
-            self.logger.exception(msg)
-            sys.exit(1)
-
-    def execute_agent(
-        self,
-    ) -> None:
-        """Execute the agent.
-        - args: background (bool): Run the agent in the background.
-        """
-        self.logger.info("Starting agent execution...")
-        try:
-            result = self.execute_command("aea -s run", verbose=True)
-            if result:
-                self.logger.info("Agent execution completed successfully. 😎")
-            else:
-                self.logger.error("Agent execution failed.")
-                sys.exit(1)
-        except RuntimeError as error:
-            self.logger.exception(f"Agent ended with error: {error}")
-        self.logger.info("Agent execution complete. 😎")
-
-    def execute_command(self, command: str, verbose=None, env_vars=None) -> None:
-        """Execute a shell command."""
-        current_vars = deepcopy(os.environ)
-        if env_vars:
-            current_vars.update(env_vars)
-        cli_executor = CommandExecutor(command=command.split(" "))
-        result = cli_executor.execute(stream=True, verbose=verbose, env_vars=current_vars)
-        if not result:
-            self.logger.error(f"Command failed: {command}")
-            self.logger.error(f"Error: {cli_executor.stderr}")
-            msg = f"Command failed: {command}"
-            raise RuntimeError(msg)
-        return result
-
-    def get_version(self) -> str:
-        """Get the version of the agent."""
-        agent_config = load_autonolas_yaml(PackageType.AGENT, self.agent_dir)[0]
-        return agent_config["version"]
+@cli.group()
+def run() -> None:
+    """Command group for running agents either in development mode or in production mode."""
 
 
-@cli.command()
+@run.command()
 @click.argument(
     "agent_public_id",
     type=PublicId.from_str,
@@ -315,13 +34,14 @@ class AgentRunner:
 @click.option("--force", is_flag=True, help="Force overwrite of existing agent", default=False)
 @click.option("--fetch/--no-fetch", help="Fetch from registry or use local agent package", default=True)
 @click.pass_context
-def run(ctx, agent_public_id: PublicId, verbose: bool, force: bool, fetch: bool) -> None:
+def dev(ctx, agent_public_id: PublicId, verbose: bool, force: bool, fetch: bool) -> None:
     """Run an agent from the local packages registry or a local path.
 
     Example usage:
-        adev run eightballer/my_agent  # Fetch and run from registry
-        adev run eightballer/my_agent --no-fetch  # Run local agent package named my_agent
+        adev run dev eightballer/my_agent  # Fetch and run from registry
+        adev run dev eightballer/my_agent --no-fetch  # Run local agent package named my_agent
     """
+
     if not agent_public_id:
         # We set fetch to false if the agent is not provided, as we assume the user wants to run the agent locally.
         fetch = False
@@ -329,12 +49,57 @@ def run(ctx, agent_public_id: PublicId, verbose: bool, force: bool, fetch: bool)
         agent_public_id = PublicId.from_json(agent_config)
     logger = ctx.obj["LOGGER"]
 
-    runner = AgentRunner(
+    runner = DevAgentRunner(
         agent_name=agent_public_id,
         verbose=verbose,
         force=force,
         logger=logger,
         fetch=fetch,
+    )
+    runner.run()
+    logger.info("Agent run complete. 😎")
+
+
+@run.command()
+@click.argument(
+    "service_public_id",
+    type=PublicId.from_str,
+    required=False,
+)
+@click.option("-v", "--verbose", is_flag=True, help="Verbose mode.", default=False)
+@click.option("--force/--no-force", is_flag=True, help="Force overwrite of existing service", default=False)
+@click.option("--fetch/--no-fetch", help="Fetch from registry or use local service package", default=True)
+@click.option("--keysfile", help="Path to the private keys file.", type=click.File(), default="keys.json")
+@click.option("--number_of_agents", "-n", help="Number of agents to run.", type=int, default=1)
+@click.pass_context
+def prod(
+    ctx,
+    service_public_id: PublicId,
+    verbose: bool,
+    force: bool,
+    fetch: bool,
+    keysfile: click.File,
+    number_of_agents: int,
+) -> None:
+    """Run an agent in production mode.
+
+    Example usage:
+        adev run prod eightballer/my_service
+    """
+
+    logger = ctx.obj["LOGGER"]
+    if not Path(keysfile.name).exists():
+        logger.error(f"Keys file not found at {keysfile.name}")
+        sys.exit(1)
+
+    runner = ProdAgentRunner(
+        service_public_id=service_public_id,
+        verbose=verbose,
+        logger=logger,
+        force=force,
+        fetch=fetch,
+        keysfile=Path(keysfile.name).absolute(),
+        number_of_agents=number_of_agents,
     )
     runner.run()
     logger.info("Agent run complete. 😎")
