+++ auto_dev/data/repo/templates/autonomy/adev_config.yaml
@@ -39,4 +39,5 @@ poetry_dependencies:
   version: 0.2.90
   location: remote
   url: https://api.github.com/repos/8ball030/auto_dev
-  extras: all
+  extras: 
+  - all
