+++ auto_dev/services/runner/__init__.py
@@ -1,0 +1,8 @@
+"""Runner service package."""
+
+from .runner import DevAgentRunner
+from .prod_runner import ProdAgentRunner
+
+
+DevAgentRunner = DevAgentRunner
+ProdAgentRunner = ProdAgentRunner
